pyproject.toml
[tool.poetry]
name = "dot"
version = "3.0.0"
description = "My dotfiles"
authors = ["Tucker Beck <tucker.beck@gmail.com>"]
license = "MIT"
packages = [ { include = "dot_tools" } ]

[tool.poetry.dependencies]
python = "^3.8"
GitPython = "^3.1.8"
pendulum = "^2.1.2"
bidict = "^0.21.2"
click = "^7.1.2"
git-url-parse = "^1.2.2"
inflection = "^0.5.1"
loguru = "^0.5.3"
pprintpp = "^0.4.0"
py-buzz = "^2.0.0"
pydon = "^0.1.0"
pytest = "^6.1.1"
requests = "^2.24.0"
sh = "^1.14.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
duck = 'bin.duck:main'
configure-dot = 'bin.configure_dot:main'
checkout-branch-by-pattern = 'bin.checkout_branch_by_pattern:main'
current-branch = 'bin.current_branch:main'
find-source-file = 'bin.find_source_file:main'
find-test-file = 'bin.find_test_file:main'
gittop = 'bin.gittop:main'
make-branch = 'bin.make_branch:main'
now = 'bin.now:main'
sep-bar = 'bin.sep_bar:main'
tag-version = 'bin.tag_version:main'
transpose = 'bin.transpose:main'
underlined-header = 'bin.underlined_header:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
